Follow the stupid Arduino naming /requirements/
The arduino environment /requires/ that you have your sketch in a file
named the same as the directory it is in.  The arduino-cli build tool
*refuses* to even consider compiling without that and the arduino GUI
simply opens an empty file with the "right" name as well as any file you
open.

This makes decades of programming tools, utilities and assumptions just
break and makes naming things even harder than it needs to be.  If I
give this repo a name that makes sense for github, it means the filename
inside it must make less sense for every one else (and the reverse, etc)

In an attempt to make this more seamless for arduino users, kowtow to
their horrid requirements and submit to their all knowing
non-overridable opinionated bullshit.