Add recipe for downloading required arduino core
The arduino-bin needs to know what FQBN to use to compile.  Each FQBN in
turn clearly shows the name of the core to use.  The list of avilable
cores is data that arduino-bin knows how to fetch.  The arduino-bin can
automatically refresh the cache of cores and download the correct core
files.

Sadly, even though it has all this information at build time, it doesnt
have an option to just-fucking-do-it, so we need to have another target
that does the work that arduino-cli could do for us.  Its like this dev
environment has never been used in an automated way.