Attempt to use known software versions
This attempt is doomed to fail, since the arduino-cli will automatically
download a bunch of other packages - some of which could easily affect
the build environment and thus the repeatability of the project.

Sadly, there seems to be no way to list or lock any of these package
versions with the arduino-cli.  Its like the people writing it have
never had concerns about supply chain verification.